Move pkg/kubectl/util to staging
