+++ pkg/kubectl/generate/versioned/configmap.go
@@ -27,9 +27,9 @@ import (
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/validation"
+	"k8s.io/kubectl/pkg/util"
 	"k8s.io/kubectl/pkg/util/hash"
 	"k8s.io/kubernetes/pkg/kubectl/generate"
-	"k8s.io/kubernetes/pkg/kubectl/util"
 )
 
 // ConfigMapGeneratorV1 supports stable generation of a configMap.
