+++ pkg/kubectl/proxy/BUILD
@@ -21,11 +21,11 @@ go_library(
     srcs = ["proxy_server.go"],
     importpath = "k8s.io/kubernetes/pkg/kubectl/proxy",
     deps = [
-        "//pkg/kubectl/util:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/net:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/proxy:go_default_library",
         "//staging/src/k8s.io/client-go/rest:go_default_library",
         "//staging/src/k8s.io/client-go/transport:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/util:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
 )
