+++ pkg/kubectl/rolling_updater_test.go
@@ -41,7 +41,7 @@ import (
 	manualfake "k8s.io/client-go/rest/fake"
 	testcore "k8s.io/client-go/testing"
 	"k8s.io/kubectl/pkg/scheme"
-	"k8s.io/kubernetes/pkg/kubectl/util"
+	"k8s.io/kubectl/pkg/util"
 )
 
 func oldRc(replicas int, original int) *corev1.ReplicationController {
