+++ staging/src/k8s.io/kubectl/BUILD
@@ -11,22 +11,7 @@ filegroup(
         ":package-srcs",
         "//staging/src/k8s.io/kubectl/pkg/rawhttp:all-srcs",
         "//staging/src/k8s.io/kubectl/pkg/scheme:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/certificate:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/deployment:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/event:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/fieldpath:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/hash:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/interrupt:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/logs:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/podutils:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/printers:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/qos:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/rbac:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/resource:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/slice:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/storage:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/templates:all-srcs",
-        "//staging/src/k8s.io/kubectl/pkg/util/term:all-srcs",
+        "//staging/src/k8s.io/kubectl/pkg/util:all-srcs",
         "//staging/src/k8s.io/kubectl/pkg/validation:all-srcs",
     ],
     tags = ["automanaged"],
